backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
groq==0.11.0
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.9
httpx==0.27.0
